* commit 'a6fa2675163aa5bd4fc4e5d70e31541f5bc50e23':
  Fix UT
  REL-306247 Call InitializeExportAsync with proper parameters
  REL-306247 Call InitializeExportAsync with proper parameters

+++ Source/Relativity.Sync.Tests.Integration/Stubs/ConfigurationStub.cs
@@ -17,7 +17,7 @@ namespace Relativity.Sync.Tests.Integration.Stubs
 		public string FieldMappings { get; set; }
 		public bool IsSnapshotCreated { get; set; }
 		
-		public async Task SetSnapshotDataAsync(Guid runId, int totalRecordsCount)
+		public async Task SetSnapshotDataAsync(Guid runId, long totalRecordsCount)
 		{
 			await Task.Yield();
 		}
