+++ Source/Relativity.Sync.Tests.System/ConfigurationRdoTests.cs
@@ -106,7 +106,7 @@ namespace Relativity.Sync.Tests.System
 			configuration.GetFieldValue<int>(RdoArtifactTypeIdGuid);
 			configuration.GetFieldValue<int>(SavedSearchInDestinationArtifactIdGuid);
 			configuration.GetFieldValue<string>(SnapshotIdGuid);
-			configuration.GetFieldValue<int>(SnapshotRecordsCountGuid);
+			configuration.GetFieldValue<long>(SnapshotRecordsCountGuid);
 			configuration.GetFieldValue<int>(SourceJobTagArtifactIdGuid);
 			configuration.GetFieldValue<string>(SourceJobTagNameGuid);
 			configuration.GetFieldValue<int>(SourceWorkspaceTagArtifactIdGuid);
@@ -288,7 +288,7 @@ namespace Relativity.Sync.Tests.System
 				{
 					Guid = SnapshotRecordsCountGuid
 				},
-				Value = 813794
+				Value = 813794L
 			};
 			yield return new FieldRefValuePair
 			{
