+++ Source/Relativity.Sync/Configuration/IDataSourceSnapshotConfiguration.cs
@@ -16,6 +16,6 @@ namespace Relativity.Sync.Configuration
 
 		bool IsSnapshotCreated { get; }
 
-		Task SetSnapshotDataAsync(Guid runId, int totalRecordsCount);
+		Task SetSnapshotDataAsync(Guid runId, long totalRecordsCount);
 	}
 }
